requirements.txt
assemblyai
python-dotenv
moviepy